+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 CHANGELOG
 =========
 
+Next Version
+============
+
+* Changed the preferred way to do model registration via model inheritance
+  and ``mptt.AlreadyRegistered``, which is deprecated since django-mptt 0.4
+
+
 1.7.1 (2020-04-29)
 ==================