+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ Next Version
 
 * Changed the preferred way to do model registration via model inheritance
   and ``mptt.AlreadyRegistered``, which is deprecated since django-mptt 0.4
-
+* Add parameter ``on_delete`` into ``FilerFolderField.__init__`` for django>=1.9.
 
 1.7.1 (2020-04-29)
 ==================